Fix delete bookmark and renew jwt errors
This patch fixes the error resulting from test migration tool change.

Previously, the deletion bookmark test assumes the target bookmark ID
to 1 which depended on the table schema. By specifying the bookmark ID
in the creation model, the dependecy is removed.

Also, fix the renewJwt test of the usage of the incorrect field(ID)
to refer the user ID(UserID).

Last, refactor these tests info table-driven tests form for readiblity.