Remove vague bookmark update
This pathc removes the vague bookmark update during creation. As of
previous bookmark schema design, the content of the post snippet (e.g.,
title) was replicated in the bookmarks table, which might be
inconsistent with original post. However, this should be addressed in
another way without updating the context consistently whenever the
bookmark is performed by a user.