api/news: implement posts query filter
This patch implments the filters parsed from the query
parameter(category, tag, id). Note that this is a breaking change which
adapted from v1 which only accepts json string to be parsed as the
mongo query document directly. However, previous implementation is not
intuitive in url context and storage dependent. This patch changes value
to use the specific parameters and transform the parameters into
mongo query respectively.